hw2/part2/requirements.txt
pytorch-pretrained-bert==0.6.1
torch==1.0
spacy==2.1.3
ftfy==4.4.3
tqdm
python-box
ipdb
pandas
PyYAML
git+https://github.com/ildoonet/pytorch-gradual-warmup-lr.git@master
